requirements.txt
Pillow>=10.0.0
requests>=2.31.0
boto3>=1.34.0
python-dotenv>=1.0.0
PyMuPDF>=1.24.2
python-barcode>=0.15.1 